Added functionality
Added thumb detection functionality. The program won't look for emotional detection until it detects a thumbs up, then it will look for an expressed emotion. Once the emotion is detected and the mirrored emotion video is played, it is reset, so the user has to do a thumbs up again in order to detect an emotion again.