Updates
Commented out the mouse movement. If using one monitor only, will need to uncomment so that the mouse will be moved to the top left of the screen at the beginning of the program and out of the way of the videos. If using two monitors, not necessary to move the mouse.

Removed "thumbsUp.py" file as it was for testing and not used in the final program.